requirements.txt
django==2.2.13
gunicorn==20.0.4
djangorestframework==3.11.0
drf-tracking==1.5.0
dj-database-url==0.5.0
python-decouple==3.3
requests==2.22.0
django-heroku==0.3.1
django-cors-headers==3.4.0
Pillow==7.2.0
